+++ stories/manifold-data-product-logo.stories.js
@@ -5,5 +5,11 @@ storiesOf('Product Logo [Data]', module)
   .addParameters({ readme: { sidebar: markdown } })
   .add(
     'default',
-    () => '<manifold-data-product-logo product-label="aiven-redis"></manifold-data-product-logo>'
+    () =>
+      `
+        <manifold-connection>
+          <manifold-auth-token token="empty"/>
+          <manifold-data-product-logo product-label="aiven-redis"></manifold-data-product-logo>
+        </manifold-connection>
+      `
   );